refactor: adopt the ui kit kbd for the header keycaps
kbd shipped only in the radix tree with no consumers; mirror it into base, tune its default look to the keycap treatment the headers use, and swap the docs keycap markup for Kbd and KbdGroup from the kit.
